+++ main/system.h
@@ -21,7 +21,7 @@ typedef struct
     uint16_t shares_rejected;
     int screen_page;
     char oled_buf[20];
-    uint32_t best_nonce_diff;
+    uint64_t best_nonce_diff;
     char best_diff_string[DIFF_STRING_SIZE];
     bool FOUND_BLOCK;
     bool startup_done;
